+++ lib/stream-to.js
